backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
anthropic==0.34.2
python-dotenv==1.0.1
alembic==1.13.3
httpx==0.27.2
